+++ packages/adminui-vue/components/controls/UFile/buttons/UFileWebcamButton.vue
@@ -224,8 +224,6 @@
 </template>
 
 <script>
-const LS_RATIO_KEY = 'UFileWebcamButton__videoRatio'
-
 export default {
   name: 'UFileWebcamButton',
   inject: {
@@ -309,12 +307,17 @@ export default {
 
     isPictureTaken () {
       return this.previewImageSrc !== null
+    },
+
+    videoRatioStorageKey () {
+      return this.$uiSettingsStorage.getKey('UFileWebcamButton', 'videoRatio')
     }
   },
 
   mounted () {
+    const videoRatioFromStorage = this.$uiSettingsStorage.getItem(this.videoRatioStorageKey)
     this.fullScreen = this.startFullScreen
-    this.videoRatio = JSON.parse(window.localStorage.getItem(LS_RATIO_KEY)) || this.videoRatios[0]
+    this.videoRatio = videoRatioFromStorage ? JSON.parse(videoRatioFromStorage) : this.videoRatios[0]
   },
 
   methods: {
@@ -544,7 +547,7 @@ export default {
 
     clearForm () {
       this.stopStream()
-      window.localStorage.setItem(LS_RATIO_KEY, JSON.stringify(this.videoRatio))
+      this.$uiSettingsStorage.setItem(this.videoRatioStorageKey, JSON.stringify(this.videoRatio))
       this.pages = []
       this.previewImageSrc = null
       this.canvas = null